feat: major refactor - multiplayer, auth, new games, CI/CD improvements
- Add multiplayer support (lobby, turn-based games)
- Add Microsoft OAuth authentication with dev bypass
- Add PoDropSquare game and high scores
- Refactor snake game engine to separate service
- Add Toast notifications, DemoMode, SinglePlayer pages
- Fix leaderboard endpoints and integration tests
- Update deploy.yml with health check and smoke tests
- Update .gitignore: exclude publish-api/, dist/, root playwright config
- Remove orphaned src/PoDropSquare and src/PoRaceRagdoll directories
- Update infra bicep to use PoShared resources
- Add docs: Architecture_SIMPLE, DataModel_SIMPLE, SystemFlow diagrams

+++ src/PoMiniGames/PoMiniGames/Models/SnakeHighScore.cs
@@ -6,7 +6,7 @@ public sealed record SnakeHighScore
     public string Initials { get; init; } = string.Empty;
     public int Score { get; init; }
     public string Date { get; init; } = string.Empty;
-    public float GameDuration { get; init; } = 30f;
+    public double GameDuration { get; init; } = 30d;
     public int SnakeLength { get; init; }
     public int FoodEaten { get; init; }
 }
